backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
openai>=1.3.0
pandas>=2.1.1
numpy>=1.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.8.5
asyncio>=3.4.3
scikit-learn>=1.3.0
matplotlib>=3.8.0
openpyxl>=3.1.2  # For Excel file support
jinja2>=3.1.2  # For FastAPI templates
typing-extensions>=4.8.0 
typing